refactor: reorganize content types and update collection names for components
