--- server/components/about-section/schema.json
@@ -1,17 +1,0 @@
-{
-    "collectionName": "about_section_plugins",
-    "info": {
-        "name": "About Section",
-        "description": "Schema for the About Section plugin"
-    },
-    "options": {
-        "draftAndPublish": true
-    },
-    "attributes": {
-        "title": {
-            "type": "string",
-            "required": true,
-            "unique": false
-        }
-    }
-}
