+++ server/components/description-item/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_description_items",
+  "collectionName": "description_items_component_p",
   "info": {
     "displayName": "Description Item",
     "icon": "align-left",
