+++ server/components/description-section/schema.json
@@ -1,5 +1,5 @@
 {
-    "collectionName": "description_section_p",
+    "collectionName": "description_section_component_p",
     "info": {
       "displayName": "Description section",
       "icon": "info",
