+++ server/components/dropdown-field/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_secciones_generales_dropdown_fields",
+  "collectionName": "dropdown_fields_component_p",
   "info": {
     "displayName": "Campo Desplegable",
     "icon": "chevron-down",
