+++ server/components/form-section/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_secciones_generales_form_sections",
+  "collectionName": "form_section_component_p",
   "info": {
     "displayName": "Sección de Formulario",
     "icon": "write",
