--- server/components/help-section/schema.json
@@ -1,38 +1,0 @@
-{
-  "collectionName": "components_help_sections",
-  "info": {
-    "displayName": "Help Section",
-    "icon": "question-circle",
-    "description": "Sección de ayuda con elementos"
-  },
-  "options": {},
-  "attributes": {
-    "title": {
-      "type": "string",
-      "required": true,
-      "maxLength": 255,
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      }
-    },
-    "subtitle": {
-      "type": "text",
-      "required": false,
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      }
-    },
-    "items": {
-      "type": "json",
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      }
-    }
-  }
-}
