+++ server/components/hero-section/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_hero_sections",
+  "collectionName": "hero_section_component_p",
   "info": {
     "displayName": "seccion de Hero (PluginComponent)",
     "icon": "image",
