+++ server/components/option/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_secciones_generales_options",
+  "collectionName": "option_component_p",
   "info": {
     "displayName": "Opción",
     "icon": "tag",
