--- server/components/sitemap-section/schema.json
@@ -1,28 +1,0 @@
-{
-  "collectionName": "components_sitemap_sections",
-  "info": {
-    "displayName": "Sitemap Section",
-    "icon": "sitemap",
-    "description": "Sección de mapa del sitio"
-  },
-  "options": {},
-  "attributes": {
-    "content": {
-      "type": "blocks",
-      "required": true,
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      }
-    },
-    "sections": {
-      "type": "json",
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      }
-    }
-  }
-}
