+++ server/components/text-field/schema.json
@@ -1,5 +1,5 @@
 {
-  "collectionName": "components_secciones_generales_text_fields",
+  "collectionName": "text_field_component_p",
   "info": {
     "displayName": "Campo de Texto",
     "icon": "pencil",
