+++ server/content-types/index.ts
@@ -1,9 +1,7 @@
-import item from './item';
 import generalPage from './general-page';
-import pageStructure from './page-structure';
+import visualType from './visual-type';
 
 export default {
-  item: { schema: item },
   'general-page': { schema: generalPage },
-  'page-structure': { schema: pageStructure },
+  'visual-type': { schema: visualType },
 };
