--- server/content-types/page-structure/index.ts
@@ -1,3 +1,0 @@
-import schema from './schema.json';
-
-export default schema;
