+++ server/content-types/visual-type/index.ts
