+++ server/content-types/visual-type/schema.json
@@ -1,0 +1,48 @@
+{
+  "kind": "collectionType",
+  "collectionName": "visual_type_type_p",
+  "info": {
+    "singularName": "visual-type",
+    "pluralName": "visual-types",
+    "displayName": "Tipo Visual(PluginType)",
+    "description": "Tipos de estilo visual para las páginas"
+  },
+  "options": {
+    "draftAndPublish": true
+  },
+  "pluginOptions": {
+    "content-manager": {
+      "visible": true
+    },
+    "content-type-builder": {
+      "visible": true
+    }
+  },
+  "attributes": {
+    "title": {
+      "type": "string",
+      "required": true,
+      "maxLength": 255,
+      "pluginOptions": {
+        "i18n": {
+          "localized": true
+        }
+      }
+    },
+    "description": {
+      "type": "text",
+      "required": true,
+      "pluginOptions": {
+        "i18n": {
+          "localized": true
+        }
+      }
+    },
+    "generalPages": {
+      "type": "relation",
+      "relation": "oneToMany",
+      "target": "plugin::secciones-generales.general-page",
+      "mappedBy": "visualType"
+    }
+  }
+}
