+++ server/register.ts
@@ -1,10 +1,4 @@
 import { Strapi } from '@strapi/strapi';
-
-// Importa los schemas de los componentes
-import aboutSectionSchema from './components/about-section/schema.json';
-import creditsSectionSchema from './components/credits-section/schema.json';
-import helpSectionSchema from './components/help-section/schema.json';
-import sitemapSectionSchema from './components/sitemap-section/schema.json';
 import heroSectionSchema from './components/hero-section/schema.json';
 import descriptionsCollectionSchema from './components/DescriptionsCollection/schema.json';
 import descriptionItemSchema from './components/description-item/schema.json';
@@ -15,6 +9,8 @@ import formSectionSchema from './components/form-section/schema.json';
 import dropdownFieldSchema from './components/dropdown-field/schema.json';
 import textFieldSchema from './components/text-field/schema.json';
 import optionSchema from './components/option/schema.json';
+import link from './components/link/schema.json';
+import redirectSection from './components/redirect-section/schema.json';
 
 // Función para convertir kebab-case a PascalCase
 const toPascalCase = (str: string) => {
@@ -29,10 +25,6 @@ export default ({ strapi }: { strapi: Strapi }) => {
   const category = 'secciones-generales';
 
   const componentsToRegister = [
-    { name: 'about-section', schema: aboutSectionSchema },
-    { name: 'credits-section', schema: creditsSectionSchema },
-    { name: 'help-section', schema: helpSectionSchema },
-    { name: 'sitemap-section', schema: sitemapSectionSchema },
     { name: 'hero-section', schema: heroSectionSchema },
     { name: 'descriptions-collection', schema: descriptionsCollectionSchema },
     { name: 'description-item', schema: descriptionItemSchema },
@@ -43,6 +35,8 @@ export default ({ strapi }: { strapi: Strapi }) => {
     { name: 'dropdown-field', schema: dropdownFieldSchema },
     { name: 'text-field', schema: textFieldSchema },
     { name: 'option', schema: optionSchema },
+    { name: 'link', schema: link },
+    { name: 'redirect-section', schema: redirectSection },
   ];
 
   componentsToRegister.forEach(({ name, schema }) => {
